create test fixture
